Spyder 4 compatibility fixes
This also removes all workarounds to keep compatibility with
Spyder 3, so the minimum version required is now v4.

As an (unwanted) side-effect, this commit removes the possibility
to pick notebooks using the file switcher. The API for this has
changed and a future commit will have to re-introduce the switcher.