Plugin: Remove .move_tab()
The code suggests that this function is called when the user moves a tab
in order to keep the list in `self.clients` in the same order as the tabs.
However, as far as I can tell, this function is never actually called,
the order in `self.clients` is not synchronized, and the order is not
actually used.