Utils: Use new ServerManager class for managing notebook servers.
This is a complete rewrite on how the plugin starts notebook servers.
A new ServerManager class starts the servers and also monitors them
and shuts them down. It uses QProcess instead of subprocess for
better control. Servers are now started asynchronously; the new
class uses signals to communicate when a server is ready. The program
now has to explicitly request the servers to shut down at the end.

A consequence of rewriting from scratch is that this is a large
commit. However, an added benefit is that we avoid potential
license issues with nbopen.