Client: Check status when retrieving sessions before parsing reply
Having received the response from the server to our request to
list all sessions, first check whether the status is ok and
only if it is, parse the reply. Before, this was done in the
opposite order.

Update the test to ensure reply is not parsed if status is
not ok.