Widget: Wait a bit repeatedly after save before reading notebook.
When saving a newly created notebook in a temporary directory, we read it
back to check whether it is empty; if not, we offer to save it in a more
permanent location. Before, we waited one second between initiating the
save and reading it back. Now, we wait 0.25 second, but keep trying if
the file does not yet exist. This means the function runs faster if the
save is expedient and it waits longer if the save runs slowly.