Widget: Display page in external browser when link in notebook is clicked
When the user clicks on a link in a notebook, the NotebookWidget calls
createWindow() in order to open a new tab to display the link. This commit
intercepts the call and creates a WebViewInBrowser object. When the URL in
that object is set, the link is opened in an external web browser.