Widgets: Use interpreter preference when opening new tab
Record the Python interpreter when a notebook server is started. When a new
tab is opened, only use an existing server if it uses the interpreter that
is currently set in the preferences, start a new server otherwise. The effect
is that if the user changes the Python interpreter, new notebooks will use the
new interpreter.