Specify file for storing connection info when starting server
Add a command line parameter to the notebook server for specifying the
file in which connection information is to be stored. Use this parameter
to set the file name to spynbserver-NNNN-MM.json where NNNN is the PID of
the Spyder process and MM distinguishes between the notebook servers started
by Spyder.

The reason for this is that Jupyter by default uses nbserver-XXXX.json
where XXXX is the PID of the notebook server process, but in a venv in
Windows there is no good method for finding this PID.