Pass connection file when opening console
Before, we passed the kernel ID when calling `create_client_for_kernel()` in the IPython
Console plugin. This worked, even though that function expected the connection file, because
that function used `find_connection_file()` to fix the path to the connection file. After the
refactor in PR spyder-ide/spyder#19062, the argument passed for the connection file is
expected to be really the file path and no attempt to fix it is made, so now we have to
really pass the connection file.