Fix interrupting Spyder kernel in notebook
Interrupting a kernel is done on Unix by sending SIGINT to the
Python process. However, Spyder kernels use `conda run` to run the
kernel in the correct environment. The SIGINT signal should not be
sent to the `conda run` process, because that causes it to quit.
Instead, find the Python process started by `conda run`, because it
will handle SIGINT correctly by raising a KeyboardInterrupt.