Only target grandchild when kernel is started with conda run
The fix in the previous commit should only be applied if the
notebook kernel was started with `conda run`, in which case the
PID in the info file identifies the conda process. Otherwise,
the PID identifies the kernel process, as normal for Jupyter
kernel, and the normal implementation for interrupting should
be followed.