Ensure that only one kernel spec is available
If the user has installed other kernel specs, for instance for
IPython kernels in other environments, and then changes kernel
then all the kernel specs are shown with the same text (the
display name of our own kernel spec). This is confusing.

This commit ensures that only one kernel spec is shown.