Let Spyder know when notebook becomes dirty
Connect to the JavaScript signal emitted when a notebook becomes
dirty or non-dirty. On receiving this signal, call alert() with
a specially crafted message. Monitor alerts in Spyder and on
receiving such an alert, emit a Qt signal.

Using alert() as a side channel to communicate messages from
JavaScript to Spyder is perhaps a bit hacky but it is simple and
it works (as long as the alert message does not occur naturally).
I expect that this mechanism is more widely applicable.