Enable Save action only if notebook is dirty
Enable the File > Save menu item and the Save button in the Spyder
toolbar only if there the current notebook is dirty (i.e., if there
is something to save).

Enable the Save All action only if any notebook is dirty. If the
user triggers this action, then only save notebooks that are dirty.